prevent internal links of imported learning modules from pointing to (existing) learning modules in trash
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@3975 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ classes/class.ilObjFile.php
@@ -244,8 +244,15 @@ class ilObjFile extends ilObject
 
 			// unlink file
 			$file = $this->getDirectory()."/".$this->getFileName();
-			unlink($file);
-			rmdir($this->getDirectory());
+			if (@is_file($file))
+			{
+				unlink($file);
+			}
+
+			if (@is_dir($this->getDirectory()))
+			{
+				rmdir($this->getDirectory());
+			}
 
 			return true;
 		}
